Restyle & refactor MessageDetail dialog
