+++ scss/modules/_bp3-overwrites.scss
@@ -66,7 +66,7 @@ $bp3-dialog-radius: 2px;
 .bp3-callout {
   width: calc(100% + 40px);
   left: -20px;
-  padding: 10px 20px 9px;
+  padding: 10px 20px;
 }
 
 .bp3-heading {
