+++ scss/modules/_message-detail.scss
@@ -1,10 +1,14 @@
 // Module: Message Detail
 
 .module-message-detail {
-  max-width: 650px;
-  margin-left: auto;
-  margin-right: auto;
-  padding: 20px;
+  margin-top: -20px;
+  .bp3-callout {
+    max-height: 50vh;
+    overflow: auto;
+  }
+  p {
+    white-space: pre-line;
+  }
 }
 
 .module-message-detail__message-container {
