+++ src/renderer/components/helpers/ChatMethods.js
@@ -8,7 +8,7 @@ export function openLeaveChatDialog (screenContext, chatId) {
   const tx = window.translate
   screenContext.openDialog('ConfirmationDialog', {
     message: tx('ask_leave_group'),
-    cb: yes => { if (yes) callDcMethod('leaveGroup', chatId) }
+    cb: yes => yes && callDcMethod('leaveGroup', chatId)
   })
 }
 
@@ -26,7 +26,7 @@ export function openBlockContactDialog (screenContext, selectedChat) {
     const contact = selectedChat.contacts[0]
     screenContext.openDialog('ConfirmationDialog', {
       message: tx('ask_block_contact'),
-      cb: yes => { if (yes) callDcMethod('blockContact', contact.id) }
+      cb: yes => yes && callDcMethod('blockContact', contact.id)
     })
   }
 }
